LRO: Properly handle flows that differ only by vlan tag
If VIMAGE is in use, it is possible to have two vlan interfaces
on the same parent interface in different VNETs, and have the
same IP address configured on both interfaces.  In this case, it
is theoretically possible for two flows to have the same 5-tuple
but actually be different connections.

Handle this case in LRO by using a vlan tag as an additional
identifier for a flow.  For an untagged packet use EVL_VLID_MASK
as a dummy value for the vlan tag, as that value is not a valid
vlan tag.